+++ NEWS.md
@@ -1,4 +1,10 @@
 
+
+# weitrix 0.99.1
+
+* Switch calibration from linear model on log dispersions to using a gamma GLM.
+* Add weitrix_calibrate_all function for very flexible calibration.
+
 # weitrix 0.99.0
 
 * Bioconductor submission.